added inheritence for the Unit class to drive Samurai and Sabry from

+++ Risk of getting Railed/Assets/Scripts/BattleHUD.cs
@@ -3,21 +3,18 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class BattleHUD : MonoBehaviour
-{
+public class BattleHUD : MonoBehaviour {
 
 	public Text nameText;
 	public Slider hpSlider;
 
-	public void SetHUD(Unit unit)
-	{
+	public void SetHUD(Unit unit) {
 		nameText.text = unit.UnitName;
 		hpSlider.maxValue = unit.MaxHP;
 		hpSlider.value = unit.CurrentHP;
 	}
 
-	public void SetHP(int hp)
-	{
+	public void SetHP(float hp) {
 		hpSlider.value = hp;
 	}
